refactor: Move mobile navigation related styling to /styled/header folder
