+++ src/components/fixtures/MobileNav.tsx
@@ -23,7 +23,7 @@ import {
     MobileNavSidebarWrap,
     MobileNavMenuList,
     MobileNavMenuItem,
-} from "../styled/MobileNavigation";
+} from "../styled/Header/MobileNavigation";
 
 /**
  * MobileNavigation
